fix(ci): resolve uintptr_t build error + MISRA violations in new modules
- Dcm.c: replace (uint8*)(uintptr_t) cast with (uint8*) — uintptr_t
  not available without <stdint.h> on CI Ubuntu GCC
- Det.c: suppress MISRA 21.6 on <stdio.h> include (SIL_DIAG output)
- SchM.h: add SchM_Test_IsIrqDisabled declaration (MISRA 8.4)
- Makefile.posix: add Det.c, SchM.c, CanTp.c to MISRA source list

+++ firmware/shared/bsw/services/Dcm.c
@@ -85,7 +85,7 @@ static void dcm_send_response(const uint8* data, PduLengthType length)
         return;
     }
 
-    pdu_info.SduDataPtr = (uint8*)(uintptr_t)data;
+    pdu_info.SduDataPtr = (uint8*)data;  /* cast away const for AUTOSAR PduInfoType */
     pdu_info.SduLength  = length;
 
     if (length > 7u) {
